+++ .github/workflows/publish-tauri.yml
@@ -29,7 +29,7 @@ jobs:
           sudo apt-get update
           sudo apt-get install -y libgtk-3-dev webkit2gtk-4.0 libappindicator3-dev librsvg2-dev patchelf
       - name: install app dependencies and build it
-        run: yarn && yarn build
+        run: yarn && VITE_FRONTEND_VERSION=__VERSION__ yarn build
       - uses: tauri-apps/tauri-action@v0
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
