+++ src-tauri/Cargo.toml
@@ -18,7 +18,6 @@ tauri-build = { version = "1.0.4", features = [] }
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 tauri = { version = "1.0.5", features = ["api-all"] }
-tauri-plugin-window-state = "0.1"
 
 [features]
 # by default Tauri runs in production mode
