+++ src/views/NotePreviewer/style.ts
@@ -93,6 +93,10 @@ a {
     font-size: 1rem;
   }
 
+  strong {
+    font-weight: bold;
+  }
+
   blockquote {
     margin: 0 0 1.5rem 0;
     padding: 0.5rem 1.5rem;
