+++ tsconfig.json
@@ -1,6 +1,7 @@
 {
   "compilerOptions": {
     "target": "es5",
+    "types": ["vite/client"],
     "downlevelIteration": true,
     "lib": [
       "dom",
